refactor: 상수를 각기 별도의 모듈로 관리하도록 분리
