+++ src/constants/auth.ts
@@ -1,0 +1,3 @@
+const username = 'a@a.com';
+const password = '1234';
+export const base64 = btoa(username + ':' + password);
