--- src/constants/index.ts
@@ -1,8 +1,0 @@
-export const CART_LOCAL_STORAGE_KEY = 'cart';
-
-export const PRODUCTS_BASE_URL = '/products';
-export const CART_BASE_URL = '/cart-items';
-
-const username = 'a@a.com';
-const password = '1234';
-export const base64 = btoa(username + ':' + password);
